Moved DeserializationException
